Moved models to their own folder and broke out correlation to it's own file.
